test: resolve c06 runtime api review findings
