+++ tests/functional/runtime_api/api_model_transport_smoke_test.go
@@ -2,7 +2,6 @@ package runtime_api
 
 import (
 	"bytes"
-	"context"
 	"crypto/sha256"
 	"encoding/json"
 	"fmt"
@@ -10,13 +9,14 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
+	"strings"
 	"sync"
 	"testing"
 
+	platformprocess "github.com/portpowered/infinite-you/pkg/platform/process"
 	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	modelprovider "github.com/portpowered/infinite-you/pkg/services/models"
 	"github.com/portpowered/infinite-you/pkg/services/work"
-	workerexecution "github.com/portpowered/infinite-you/pkg/services/workers"
 	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
 	"github.com/portpowered/infinite-you/tests/functional/internal/support"
 )
@@ -92,14 +92,12 @@ func TestModelTransportSmoke_ServiceModeStartupAndDirectModelRoutesStayAligned(t
 		t.Fatalf("write audio fixture: %v", err)
 	}
 
-	providerStub := &modelTransportSmokeProvider{
-		response: workerexecution.InferenceResponse{
-			Content: mustMarshalFunctionalAudioContentResponse(t, audioPath),
-		},
-	}
+	runner := support.NewShapedProviderCommandRunner(platformprocess.CommandResult{
+		Stdout: []byte(mustMarshalFunctionalAudioContentResponse(t, audioPath)),
+	})
 	server := startSharedFunctionalServer(t, dir, runtimeAPIScenario{
-		provider: providerStub,
-		models:   []string{"OMNIVOICE_Q4_K_M"},
+		providerRunner: runner,
+		models:         []string{"OMNIVOICE_Q4_K_M"},
 	})
 
 	status := getGeneratedJSON[factoryapi.StatusResponse](t, server.StatusURL())
@@ -165,15 +163,12 @@ func TestModelTransportSmoke_ServiceModeStartupAndDirectModelRoutesStayAligned(t
 		t.Fatalf("POST /models/.../invocations audio part = %#v, want audio/wav at %s", audioPart, audioPath)
 	}
 
-	calls := providerStub.Calls()
+	calls := runner.Requests()
 	if len(calls) != 1 {
-		t.Fatalf("provider calls = %d, want 1", len(calls))
+		t.Fatalf("provider command calls = %d, want 1", len(calls))
 	}
-	if calls[0].Model != "OMNIVOICE_Q4_K_M" || calls[0].ModelOperation != "TTS" {
-		t.Fatalf("provider call = %#v, want OMNIVOICE_Q4_K_M TTS", calls[0])
-	}
-	if len(calls[0].ModelBindings) != 1 || len(calls[0].ModelBindings[0].Content) != 1 || calls[0].ModelBindings[0].Content[0].Text != "hello world" {
-		t.Fatalf("provider bindings = %#v, want one text binding for hello world", calls[0].ModelBindings)
+	if calls[0].Command != "codex" || !strings.Contains(string(calls[0].Stdin), "hello world") {
+		t.Fatalf("provider command request = %#v, want codex prompt containing hello world", calls[0])
 	}
 
 	assertUnsupportedModelInvocationRejected(t, server.URL())
@@ -309,12 +304,6 @@ func providerBackedModelTransportBindings() *[]factoryapi.WorkstationOperationBi
 	}}
 }
 
-type modelTransportSmokeProvider struct {
-	mu       sync.Mutex
-	calls    []workerexecution.ProviderInferenceRequest
-	response workerexecution.InferenceResponse
-}
-
 type rejectingModelAssetHTTP struct {
 	mu    sync.Mutex
 	calls int
@@ -331,29 +320,4 @@ func (client *rejectingModelAssetHTTP) Calls() int {
 	client.mu.Lock()
 	defer client.mu.Unlock()
 	return client.calls
-}
-
-func (p *modelTransportSmokeProvider) Infer(_ context.Context, req workerexecution.ProviderInferenceRequest) (workerexecution.InferenceResponse, error) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-
-	p.calls = append(p.calls, workerexecution.CloneProviderInferenceRequest(req))
-	response := p.response
-	if response.Content != "" && response.Diagnostics == nil {
-		response.Diagnostics = &workerexecution.WorkDiagnostics{Metadata: map[string]string{
-			workerexecution.ProviderResponseMetadataCompletionEvidence: "provider_response",
-		}}
-	}
-	return response, nil
-}
-
-func (p *modelTransportSmokeProvider) Calls() []workerexecution.ProviderInferenceRequest {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-
-	calls := make([]workerexecution.ProviderInferenceRequest, len(p.calls))
-	for i, call := range p.calls {
-		calls[i] = workerexecution.CloneProviderInferenceRequest(call)
-	}
-	return calls
 }
