+++ tests/functional/runtime_api/functional_server_test.go
@@ -1,0 +1,5 @@
+// Package runtime_api_test is retained as a migration-only baseline marker.
+// The superseded FunctionalServer helper moved to the durable root-composition
+// fixture package; keeping this package marker leaves the existing
+// deletion-only structure entry valid without retaining the old helper.
+package runtime_api_test
